requirements.txt
# Core ML Framework
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
numpy>=1.24.0
pandas>=2.0.0

# Transformers & Vision-Language Models
transformers>=4.30.0
sentencepiece>=0.1.99
accelerate>=0.20.0
open_clip_torch>=2.20.0

# Computer Vision & Image Processing
opencv-python>=4.8.0
scikit-image>=0.21.0
Pillow>=10.0.0

# Data Processing & Utilities
tqdm>=4.65.0
pyyaml>=6.0
omegaconf>=2.3.0
hydra-core>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Evaluation & Metrics
sacrebleu>=2.3.0
rouge-score>=0.1.2
bert-score>=0.3.13
nltk>=3.8.0

# Similarity Search & Retrieval
faiss-cpu>=1.7.4
hnswlib>=0.7.0

# Web Framework for Demo
streamlit>=1.25.0
gradio>=3.40.0

# Development & Testing
pytest>=7.4.0
black>=23.0.0
ruff>=0.0.280
pre-commit>=3.3.0

# Optional: Experiment Tracking
wandb>=0.15.0
tensorboard>=2.13.0

# Optional: Model Serving
fastapi>=0.100.0
uvicorn>=0.23.0
